% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_model.R
\name{malaria_model}
\alias{malaria_model}
\title{Create Malaria Model}
\usage{
malaria_model(
  params = NULL,
  EIR = NULL,
  ft = NULL,
  ton = 365,
  toff = 4015,
  day0_res = 0.01,
  init_res = 0,
  res_time = 0,
  treatment_failure_rate = 0.3,
  rT_r_cleared = 0.1,
  rT_r_failed = 0.1,
  resistance_trans_mult = 1,
  resistance_dur_mult = 1,
  resistance_baseline_ratio = 1,
  resistance_cleared_ratio = 1,
  resistance_failed_ratio = 1,
  verbose = FALSE
)
}
\arguments{
\item{params}{A list of parameters to pass to the model. If NULL, parameters will be generated based on EIR and ft.}

\item{EIR}{Entomological Inoculation Rate (used if params is NULL)}

\item{ft}{Treatment rate (used if params is NULL)}

\item{ton}{Time at which resistance is turned on}

\item{toff}{Time at which resistance is turned off}

\item{day0_res}{Resistant at Day 0. Default = 0.01}

\item{init_res}{Initial resistance level at res_time}

\item{res_time}{Time at which resistance is introduced}

\item{treatment_failure_rate}{Proportion of treatments that fail completely (default = 0.3)}

\item{rT_r_cleared}{Recovery rate for delayed clearance treatment (default = 0.1)}

\item{rT_r_failed}{Rate of transition from failed treatment to asymptomatic (default = 0.1)}

\item{resistance_trans_mult}{transmission multiplier for resistant parasites (default = 1)}

\item{resistance_dur_mult}{duration multiplier for resistant infections (default = 1)}

\item{resistance_baseline_ratio}{baseline infectiousness ratio for untreated resistant parasites (default = 1)}

\item{resistance_cleared_ratio}{infectiousness ratio for partial treatment response (default = 1)}

\item{resistance_failed_ratio}{infectiousness ratio for treatment failure (default = 1)}

\item{verbose}{Logical. If TRUE, prints detailed logs. Default is FALSE.}
}
\value{
An object of class \code{odin_model}.
}
\description{
Create Malaria Model
}
